navigate panes using Vi movement keys
Bind keys to navigate panes using Vi movement keys. There are few
different approaches to take for moving among windows and panes
similarly to how one moves among windows in Vim, as shown in the option
sections that are added to tmux.conf. The active option is a decent
compromise and is likely the preferred approach, as it keeps stays most
faithful to the default tmux bindings.