split rtl_fm and direwolf into separate services
Split rtl_fm and direwolf into separate services and stop using a shell
script to try to control the igate pipeline.

rtl_fm now runs in the aprs service with its stdout hooked up to a named
pipe and direwolf runs in the igate service with its stdin hookup to the
other end of the named pipe.