improve tmux session creation
Ensure the files sourced in bashrc are ordered. This will ensure, for
instance, that the path is set correctly when the tmux panes are
created. This is important when a command is being run directly in a
pane without first starting a login or interactive shell.

Separate the creation of the tmux server and the creation of the default
tmux session.

Fix the godoc pane to work with latest godoc by removing the analysis
flag from the godoc invocation.

Make all the dashboard panes remain on exit so that they can easily be
respawned.