bashrc: use octal escapes instead of hex escapes
Use octal escapes instead of hex escape codes, because the hex escapes
do not seem to be working correctly. With the hex escapes, the
nonprintable characters seem to be counted in line length when the lines
is beyond a certain length. Trying to use the hex codes in a string
(e.g. "\x01" instead of $'\x01') simply doesn't work and causes the
prompt to output the string literal (e.g. \x01).