better parameterize igate templates
Better parameterize the igate template to use latitude and longitude in
decimal format so that the values can be reused in other places later,
because direwolf is inconsistent in that some fields can support
multiple formats while other fields can only support decimal; use
decimal to avoid avoid having to repeat the lat/long in multiple
representations.

Separate the call from the ssid so that the call can be used in other
places, too.