developer: update git workspace and insteadOf management
Update git workspace  management to replace a commented out example with
a more generically useful workspace.

Ensure insteadOfs are at least an empty array so that the property is
not treated as a string.

Add a personal workspace.

Update the git config command used to manage insteadOfs to handle
removal of empty sections instead of having to maintain a placeholder.

Allow an insteadOf dictionary to be missing some expected properties
instead of treating all errors the same for the tasks that fail when an
expected property is missing (i.e. use when and skip_missing where
previously ignore_errors was set.